requirements.txt
amqp==5.3.1
anyio==4.9.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
Deprecated==1.2.18
Flask==3.1.1
Flask-Limiter==3.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
limits==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==25.0
prompt_toolkit==3.0.51
pycparser==2.22
Pygments==2.19.2
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.2.0
requests==2.32.4
rich==13.9.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2